+++ NEWS.md
@@ -1,5 +1,9 @@
 # tidyplots (development version)
 
+## Improvements
+
+* `add_data_points_jitter()`, `add_data_labels()` and `add_data_labels_repel()` now expose the `seed` parameter.
+
 # tidyplots 0.4.0
 
 ## Breaking changes